+++ packages/sdk/src/browserSdk.ts
@@ -31,8 +31,8 @@ import {
   BrowserSession,
   BrowserTransport,
 } from '@looker/sdk-rtl/lib/browser'
-import { Looker40SDK } from './sdk/4.0/methods'
-import { Looker31SDK } from './sdk/3.1/methods'
+import { Looker40SDK } from './4.0/methods'
+import { Looker31SDK } from './3.1/methods'
 
 export const BrowserSettings = (): IApiSettings => {
   const settings = DefaultSettings()